Unify UserClient and GatewayClient
