--- client/client-lib/src/clients/mod.rs
@@ -1,3 +1,0 @@
-pub mod gateway;
-pub mod transaction;
-pub mod user;
